myproject/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
slackclient==2.9.4
ibm-generative-ai==0.2.3